fix: RIG_LEVEL_AGC is an integer(enum) type, not float.
git-svn-id: https://hamlib.svn.sourceforge.net/svnroot/hamlib/trunk@1747 7ae35d74-ebe9-4afe-98af-79ac388436b8

+++ tentec/tentec.h
@@ -2,7 +2,7 @@
  *  Hamlib Tentec backend - main header
  *  Copyright (c) 2001-2004 by Stephane Fillod
  *
- *	$Id: tentec.h,v 1.8 2004-05-03 22:34:14 fillods Exp $
+ *	$Id: tentec.h,v 1.9 2004-05-26 21:30:13 fillods Exp $
  *
  *   This library is free software; you can redistribute it and/or modify
  *   it under the terms of the GNU Library General Public License as
@@ -32,7 +32,7 @@ struct tentec_priv_data {
 	int cwbfo;		/* BFO frequency: 0 [0-2000Hz] */
 	float lnvol;		/* line-out volume: 30 [0..63] */
 	float spkvol;		/* speaker volume: 30 [0..63] */
-	float agc;		/* AGC: medium*/
+	int agc;		/* AGC: medium */
 
 	/* calculated by tentec_tuning_factor_calc() */
 	int ctf; /* Coarse Tune Factor */
